gateway implement
