+++ api/api_carfile.go
@@ -18,5 +18,5 @@ type CarfileOperation interface {
 	// query block caching stat
 	QueryCachingCarfile(ctx context.Context) (*types.CachingCarfile, error) //perm:write
 	// query cache progress
-	CachedProgresses(ctx context.Context, carfileCIDs []string) ([]*types.CarfileProgress, error) //perm:write
+	CachedProgresses(ctx context.Context, carfileCIDs []string) (*types.CacheResult, error) //perm:write
 }
