+++ node/gateway/handler.go
@@ -68,7 +68,6 @@ func getEtag(r *http.Request, cid cid.Cid) string {
 		// Etag: "cid.shortFmt" (gives us nice compression together with Content-Disposition in block (raw) and car responses)
 		suffix = `.` + shortFormat + suffix
 	}
-	// TODO: include selector suffix when https://github.com/ipfs/kubo/issues/8769 lands
 	return prefix + cid.String() + suffix
 }
 
@@ -85,15 +84,9 @@ func addCacheControlHeaders(w http.ResponseWriter, r *http.Request, contentPath 
 
 	// Set Cache-Control and Last-Modified based on contentPath properties
 	if contentPath.Mutable() {
-		// mutable namespaces such as /ipns/ can't be cached forever
 
-		/* For now we set Last-Modified to Now() to leverage caching heuristics built into modern browsers:
-		 * https://github.com/ipfs/kubo/pull/8074#pullrequestreview-645196768
-		 * but we should not set it to fake values and use Cache-Control based on TTL instead */
 		modtime = time.Now()
 
-		// TODO: set Cache-Control based on TTL of IPNS/DNSLink: https://github.com/ipfs/kubo/issues/1818#issuecomment-1015849462
-		// TODO: set Last-Modified based on /ipns/ publishing timestamp?
 	} else {
 		// immutable! CACHE ALL THE THINGS, FOREVER! wolololol
 		w.Header().Set("Cache-Control", immutableCacheControl)
@@ -101,7 +94,6 @@ func addCacheControlHeaders(w http.ResponseWriter, r *http.Request, contentPath 
 		// Set modtime to 'zero time' to disable Last-Modified header (superseded by Cache-Control)
 		modtime = time.Unix(0, 0)
 
-		// TODO: set Last-Modified? - TBD - /ipfs/ modification metadata is present in unixfs 1.5 https://github.com/ipfs/kubo/issues/6920?
 	}
 
 	return modtime
